pyproject.toml
[tool.poetry]
name = "roll-call"
version = "0.1.0"
description = ""
authors = ["inth3wild"]

[tool.poetry.dependencies]
python = "^3.9"
pendulum = "^2.1.2"
requests = "^2.27.1"
pyzk = "^0.9"
termcolor = "^1.1.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
